requirements.txt
pytest==8.4.0
requests==2.32.3
pact-python==2.3.1
Flask==3.1.1
yarl==1.20.1